S08P31C104-471 fix : 이메일 인증 여부 확인 1. redis 값 조회 추가

+++ back/member-service/src/main/java/com/example/memberservice/service/MemberServiceImpl.java
@@ -101,6 +101,11 @@ public class MemberServiceImpl implements MemberService {
 
 		String encryptedPwd = bCryptPasswordEncoder.encode(signUpMemberRequestDto.getPassword());
 
+		boolean isNotAuthenticated = redisService.existsByKey(AUTH_EMAIL.getKey() + email);
+		if (isNotAuthenticated) {
+			throw new ApiException(ExceptionEnum.NOT_AUTHENTICATED_EMAIL_EXCEPTION);
+		}
+
 		redisService.deleteData(AUTH_EMAIL.getKey() + email);
 
 		memberRepository.save(signUpMemberRequestDto.toEntity(encryptedPwd));
